Finished switching the API to new return style

+++ 4coder_API/4coder_types.h
@@ -927,7 +927,7 @@ STRUCT Record_Info{
         struct{
             String_Const_u8 string_forward;
             String_Const_u8 string_backward;
-            i32 first;
+            i64 first;
         } single;
         struct{
             i32 count;
